Upload images by dragging and dropping in textareas
Images are renamed to a random UUID and stored in a directory located at
/public/assets/<namespace>/<project>/

There is no access right for images: as for note images, all images are
publicly available if the exact URL is known.

The implementation uses the 'jQuery File Upload', plus some jquery-soup.
The soup should be refactored into a coffeescript module at some point.